+++ Sources/Utils/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.30 (alpha):
 * [Enhancement] Add new method to align vessel via the nodes: [AlignTransforms.SnapAlignVessel](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_SnapAlignVessel.htm).
 * [Enhancement] Add new method to place a vessel: [AlignTransforms.PlaceVessel](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_PlaceVessel.htm).
+* [Change] Drop `traceUpdates` parameter in [`AsyncCall.WaitForPhysics`](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ProcessingUtils_AsyncCall_WaitForPhysics.htm).
 * [Fix] Handle persistent fields that are set to NULL: silently don't store them, and create on restore if there is a config node. See [`ConfigUtils.PersistentFieldAttribute`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ConfigUtils_PersistentFieldAttribute.htm).
 * [Fix] Properly calculate the rotation in [`ModelUtils.PersistentFieldAttribute.AlignTransforms.SnapAlign`](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_SnapAlign.htm).